BJS2-43978 rename package
